requirement.txt
asgiref==3.5.2 
backports.zoneinfo==0.2.1 
django==4.1.3 
sqlparse==0.4.3
Pillow==9.3.0
altgraph==0.17.3
auto-py-to-exe==2.23.1
backports.zoneinfo==0.2.1
bottle==0.12.23
bottle-websocket==0.2.9
Django==4.1.3
django-crispy-forms==1.14.0
django-moneyfield==0.2.1
djangorestframework==3.14.0
Eel==0.14.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
Pillow==9.3.0
psycopg2==2.9.5
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pyotp==2.7.0
pyparsing==3.0.9
pytz==2022.6
sqlparse==0.4.3
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.5.1